refactor: extract training spot helpers
